--- javav2/example_code/sagemaker/src/main/java/com/example/sage/ListNotebooks.java
@@ -1,58 +1,0 @@
-//snippet-sourcedescription:[ListNotebooks.java demonstrates how to list notebooks.]
-//snippet-keyword:[AWS SDK for Java v2]
-//snippet-keyword:[Amazon SageMaker]
-
-/*
-   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   SPDX-License-Identifier: Apache-2.0
-*/
-
-package com.example.sage;
-
-//snippet-start:[sagemaker.java2.list_books.import]
-import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.sagemaker.SageMakerClient;
-import software.amazon.awssdk.services.sagemaker.model.ListNotebookInstancesResponse;
-import software.amazon.awssdk.services.sagemaker.model.NotebookInstanceSummary;
-import software.amazon.awssdk.services.sagemaker.model.SageMakerException;
-import java.util.List;
-//snippet-end:[sagemaker.java2.list_books.import]
-
-/**
- * Before running this Java V2 code example, set up your development environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class ListNotebooks {
-
-    public static void main(String[] args) {
-
-        Region region = Region.US_WEST_2;
-        SageMakerClient sageMakerClient = SageMakerClient.builder()
-            .region(region)
-            .credentialsProvider(ProfileCredentialsProvider.create())
-            .build();
-
-        listBooks(sageMakerClient);
-        sageMakerClient.close();
-    }
-
-    //snippet-start:[sagemaker.java2.list_books.main]
-    public static void listBooks(SageMakerClient sageMakerClient) {
-        try {
-            ListNotebookInstancesResponse notebookInstancesResponse = sageMakerClient.listNotebookInstances();
-            List<NotebookInstanceSummary> items = notebookInstancesResponse.notebookInstances();
-            for (NotebookInstanceSummary item: items) {
-                System.out.println("The notebook name is: "+item.notebookInstanceName());
-            }
-
-        } catch (SageMakerException e) {
-            System.err.println(e.awsErrorDetails().errorMessage());
-            System.exit(1);
-        }
-    }
-    //snippet-end:[sagemaker.java2.list_books.main]
-}
